+++ packages/zero/src/components/table/index.ts
@@ -1,3 +1,3 @@
-export { Table } from './Table.js';
-export type { TableRootProps, TablePartProps, TableRowProps, TableHeaderCellProps, TableCellProps, TableCellSpanProps } from './Table.js';
+export { Table, useTableContext } from './Table.js';
+export type { TableRootProps, TablePartProps, TableRowProps, TableHeaderCellProps, TableCellProps, TableCellSpanProps, TableCellColumnProps, TableColumn } from './Table.js';
 export { tableAnatomy } from './anatomy.js';